Move RecursiveTable outside of sidebar
